Introduce abstract classes to remove duplicates
