+++ src/main/java/fr/insee/keycloak/providers/franceconnect/FranceConnectIdentityProviderConfig.java
@@ -1,17 +1,14 @@
 package fr.insee.keycloak.providers.franceconnect;
 
-import fr.insee.keycloak.providers.common.EidasLevel;
-import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
+import fr.insee.keycloak.providers.common.AbstractBaseProviderConfig;
 import org.keycloak.models.IdentityProviderModel;
 
-class FranceConnectIdentityProviderConfig extends OIDCIdentityProviderConfig {
+class FranceConnectIdentityProviderConfig extends AbstractBaseProviderConfig {
 
-  private static final EidasLevel DEFAULT_EIDAS_LEVEL = EidasLevel.EIDAS1;
   private static final FCEnvironment DEFAULT_FC_ENVIRONMENT = FCEnvironment.INTEGRATION_V1;
 
   FranceConnectIdentityProviderConfig(IdentityProviderModel identityProviderModel) {
     super(identityProviderModel);
-
     initialize();
   }
 
@@ -30,14 +27,4 @@ class FranceConnectIdentityProviderConfig extends OIDCIdentityProviderConfig {
     this.setValidateSignature(true);
     this.setBackchannelSupported(false);
   }
-
-  boolean isIgnoreAbsentStateParameterLogout() {
-    return Boolean.parseBoolean(getConfig().get("ignoreAbsentStateParameterLogout"));
-  }
-
-  EidasLevel getEidasLevel() {
-    return EidasLevel.getOrDefault(
-        getConfig().get(EidasLevel.EIDAS_LEVEL_PROPERTY_NAME), DEFAULT_EIDAS_LEVEL);
-  }
-
 }
